refactor: 에러 코드 리팩토링 적용
