+++ swith-me/swith-me-payment-api/src/main/java/project/swithme/payment/core/facade/PaymentFacade.java
@@ -1,6 +1,6 @@
 package project.swithme.payment.core.facade;
 
-import static project.study.support.codeandmessage.payment.PaymentCodeAndMessage.INVALID_PAYMENT_EXECUTION;
+import static project.study.support.codeandmessage.payment.PaymentErrorCodeAndMessage.INVALID_PAYMENT_EXECUTION;
 import java.math.BigDecimal;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
