+++ swith-me/swith-me-payment-api/src/test/java/project/swithme/payment/test/documentationtest/command/pay/PaymentFailureDocumentationTest.java
@@ -7,7 +7,7 @@ import static order.PaymentFixture.getPaymentKey;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static project.study.support.codeandmessage.payment.PaymentCodeAndMessage.INVALID_PAYMENT_EXECUTION;
+import static project.study.support.codeandmessage.payment.PaymentErrorCodeAndMessage.INVALID_PAYMENT_EXECUTION;
 import static project.swithme.domain.core.order.entity.OrderStatus.PAYMENT_REQUEST;
 import static project.swithme.payment.common.mockresponse.OrderCommandMockResponse.createNotApprovedCommand;
 import static project.swithme.payment.test.documentationtest.snippet.PaymentSnippet.TOSS_CARD_PAYMENT_FAILURE_HANDLER;
@@ -17,7 +17,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
-import org.springframework.test.web.servlet.MockMvc;
 import project.swithme.domain.core.order.entity.Order;
 import project.swithme.domain.core.payment.entity.PaymentType;
 import project.swithme.payment.common.persistence.PersistenceHelper;
